+++ controller/dlcController.js
@@ -1,9 +1,10 @@
 const Dlc = require('../model/dlcModel');
+const CustomeError = require('../utils/CustomeError');
 
 async function getDlcs(req, res) {
   try {
     const dlcs = await Dlc.find();
-    res.json(dlcs);
+    res.status(200).json({status: 200, data: dlcs});
   } catch (error) {
     res.status(500).json({ error: 'Error al obtener los DLCs' });
   }
@@ -12,8 +13,14 @@ async function getDlcs(req, res) {
 async function createDlc(req, res) {
   const { title, description, price } = req.body;
   try {
-    const dlc = await Dlc.create({ title, description, price });
-    res.status(201).json(dlc);
+    if( title!=undefined && description!=undefined && price!=undefined){
+      const dlc = await Dlc.create({ title, description, price });
+      res.status(201).json(dlc);
+    }
+    else{
+      const error = new CustomeError('Faltan parametros', 400);
+      return next(error);
+  }
   } catch (error) {
     res.status(500).json({ error: 'Error al crear el DLC' });
   }
@@ -23,14 +30,21 @@ async function updateDlcByTitle(req, res) {
   const { title } = req.params;
   const { description, price } = req.body;
   try {
-    const dlc = await Dlc.findByIdAndUpdate({ title },
-      { title, description, price },
-      { new: true }
-    );
-    if (dlc) {
-      res.json(dlc);
-    } else {
-      res.status(404).json({ error: 'DLC no encontrado' });
+    if( title!=undefined && description!=undefined && price!=undefined){
+      const dlc = await Dlc.findByIdAndUpdate({ title },
+        { title, description, price },
+        { new: true }
+      );
+      if (dlc) {
+        res.status(200).json({status: 200, data: dlc});
+      } else {
+        const error = new CustomeError('DLC no encontrado', 404);
+        return next(error);
+      }
+    }
+    else{
+      const error = new CustomeError('Faltan parametros', 400);
+      return next(error);
     }
   } catch (error) {
     res.status(500).json({ error: 'Error al actualizar el DLC' });
@@ -40,11 +54,18 @@ async function updateDlcByTitle(req, res) {
 async function deleteDlcByTitle(req, res) {
   const { title } = req.params;
   try {
-    const dlc = await Dlc.findByIdAndDelete({ title });
-    if (dlc) {
-      res.json({ message: 'DLC eliminado correctamente' });
-    } else {
-      res.status(404).json({ error: 'DLC no encontrado' });
+    if( title!=undefined ){
+      const dlc = await Dlc.findByIdAndDelete({ title });
+      if (dlc) {
+        res.status(200).json({ status: 200, message: 'DLC eliminado correctamente' });
+      } else {
+        const error = new CustomeError('DLC no encontrado', 404);
+        return next(error);
+      }
+    }
+    else{
+      const error = new CustomeError('Faltan parametros', 400);
+      return next(error);
     }
   } catch (error) {
     res.status(500).json({ error: 'Error al eliminar el DLC' });
